requirements.txt
opencv-python
face-recognition
numpy
pandas
streamlit
mysql-connector-python
pickle-mixin
openpyxl
